+++ level_editor/src/org/vertexarmy/dsr/leveleditor/editors/sprite/SpriteEditorInputComponent.java
@@ -1,5 +1,6 @@
 package org.vertexarmy.dsr.leveleditor.editors.sprite;
 
+import com.badlogic.gdx.Input;
 import com.badlogic.gdx.InputAdapter;
 import com.badlogic.gdx.InputProcessor;
 import com.badlogic.gdx.math.Vector2;
@@ -7,6 +8,7 @@ import lombok.RequiredArgsConstructor;
 import org.vertexarmy.dsr.core.DragHelper;
 import org.vertexarmy.dsr.core.component.InputComponent;
 import org.vertexarmy.dsr.core.systems.RenderSystem;
+import org.vertexarmy.dsr.math.Algorithms;
 
 /**
  * created by Alex
@@ -38,14 +40,27 @@ public class SpriteEditorInputComponent extends InputAdapter implements InputCom
     }
 
     @Override
+    public boolean keyDown(int keycode) {
+        if (keycode == Input.Keys.E) {
+            editor.getEditDialog().show();
+            return true;
+        }
+
+        return false;
+    }
+
+    @Override
     public boolean touchDown(int screenX, int screenY, int pointer, int button) {
         if (!editor.isBoundToSprite()) {
             return false;
         }
 
-        if (!editor.getRotateHandler().isHovered() && !editor.getScaleHandler().isHovered()) {
+        Vector2 mouseWorldPosition = RenderSystem.instance().screenToWorld(screenX, screenY);
+
+        if (!editor.getRotateHandler().isHovered() && !editor.getScaleHandler().isHovered() &&
+                Algorithms.createRectangle(editor.getSpriteBottomLeftCorner(), editor.getSpriteTopRightCorner()).contains(mouseWorldPosition)) {
             editMode = EditMode.MOVE;
-            dragHelper.beginDrag(RenderSystem.instance().screenToWorld(screenX, screenY));
+            dragHelper.beginDrag(mouseWorldPosition);
             return true;
         }